Data binding initialise

+++ app/src/main/java/com/attendance/letmeattend/Model/Attendance.kt
@@ -1,4 +1,4 @@
 package com.attendance.letmeattend.Model
 
-data class Attendance(var attendance: Float) {
+data class Attendance(var attendance: Int) {
 }
